Move LocalSession/Memory_Storage to import these classes for development and testing purpose.
