+++ src/main/java/in/oogway/plumbox/cli/testing/MemoryStorage.java
@@ -1,4 +1,4 @@
-package in.oogway.plumbox.cli;
+package in.oogway.plumbox.cli.testing;
 
 import in.oogway.plumbox.launcher.StorageDriver;
